upscaler-service/requirements.txt
fastapi
uvicorn[standard]
Pillow
requests
realesrgan
torch
numpy
